requirements.txt
yt-dlp>=2025.10.14
customtkinter>=5.2.0
pillow>=10.0.0
